Optimize test running a bit
When starting a test run, we set up local repository mirrors to clone when
installing Critic. This included mirrors of v8-jsshell.git and v8.git.
Since those two are only used when --test-extensions is enabled, and in
particular v8.git is fairly large, we can optimize the test start-up time
by not setting up mirrors of them when they're not needed.