Pre-start extension processes
Instead of starting extension processes as direct child processes of
whatever process triggers the extension, run a background service whose
job it is to run extension processes. Other processes connects over a
UNIX socket, sends (primarily) the input to the extension, and receives
the result.

Further, in order to optimize extension running, the background process
will maintain a "cache" of already started extension processes, that are
waiting for input on their standard input pipe.