Improve internal API caching behavior
The internal API implementation caches all objects, so that repeated
requests for an object with a certain unique identifier returns the same
object.

This is nice, but it's even nicer if we can actually use this cache to
avoid database queries altogether. This patch achieves that. The key part
is the APIObject.cached() decorator, which is designed to wrap the
internal (api.impl.*) fetch() functions. It performs a cache lookup using
the second argument as the unique identifier, and avoids calling fetch()
at all if the lookup is successful.

Commit objects are special. They have two unique identifiers (commit id
and sha1), which are now both usable as cache keys. Also, neither is
actually enough, since a repository is also required, so the cache keys
are actually (repository_id, commit_id) and (repository_id, sha1) tuples.