Split pre-receive Git hook into pre-receive + post-receive
The pre-receive hook checks that the update is valid, and if not rejects
it, and then inserts a simple record of it into the database and accepts
the update.

The new post-receive hook blocks, polling the database for update output
until the update has been fully processed. This is non-essential, if the
pushing Git client disappears, the update will still be fully processed,
and the Git repository will already have been updated.

Two new background services take care of the actual updating:

- The branchupdater service first inserts new commits into the commits
  table, calculates how the set of commits associated with the branch
  ought to be updated, and records that into the database.

- The reviewupdater service then deals with the recorded branch update
  if the branch is a review branch; creating changesets, assigns changes
  to reviewers and sends emails.

Each branch update will now be stored in the database, along with a
record of how the branch's associated set of commits was changed. This
was previously only done for review rebases, which were of course just
a special case branch update. This means that in theory all updates to
branches can be reversed, not just review rebases.