Fix time delta calculations in SQLite compatibility layer
With SQLite, all date/time values are stored as integers (seconds since
1970-01-01), so we don't need to use strftime() to interpret stored
values; only to convert "now" to a compatible integer value for doing
arithmetics on.

The behavior when calling

  strftime('%s', <seconds-since-1970>)

appears to have changed in recent versions of SQLite. It used to return
a "bogus" value (meaning this code was always quite broken) and now
returns null (which actually causes exceptions in the code.)