Support Git 2.11 and newer
In Git versions starting with 2.11, the pushed objects are not added to
the repository before running the pre-receive hook. Instead, they are
stored in a quarantine area, and GIT_ALTERNATE_OBJECT_DIRECTORIES and
GIT_OBJECT_DIRECTORY are set such that the objects can be accessed by the
pre-receive hook.

To support this, pass all GIT_* environment from the pre-receive hook to
the githook background service, and set them whenever invoking Git.