Drop the "review merge confirmation" mechanism
It was wrong and broken in all sorts of ways, and the new branch updating
machinery doesn't use it, so it's thankfully unnecessary as well.

When pushing a merge commit to a review, a simple heuristic now determine
what commits to add to the review: if a parent is a descendant of a commit
in the review, add it (and ancestors), otherwise skip it.