JSON API: Implement "internal redirect" mechanism for create
If a resource has a list of some other resource, it's typically convenient
to support creation of the other resource (with some parameters implied)
by POST:ing to the path of that list.

This is now supported by having the original resource's create() raise an
InternalRedirect exception with the resource path of the resource that
would actually be created, typically after augmenting the input somewhat.

To support the latter, add jsonapi.check.ensure() that checks that some
input field has the expected value, or sets it if the field is missing
from the input.