Set correct SYSTEM_RECIPIENTS in quick-started instance
When no --system-recipient arguments were given, the setting's value
defaulted to a string (<username> + "@localhost"). But the value should
be a *list* of email addresses, so that string was used as a sequence
of one-character email addresses, which is useless, and caused some
follow-up errors in the mail delivery system.