Improve error handling in background services
When writing to a peer, if the connection had been closed, an EPIPE error
("Broken pipe") would typically be raised. This was not handled, leading
to the whole service crashing.

Clients normally aren't expected to disconnect like that, but that doesn't
mean crashing is the appropriate way to handle it when they do.