Enhance the review api resource
Reviewers are now represented by two sets, active and assigned. A user can be
in both sets at once.

There is now a total_progress property, expressed as a number between 0 and 1
(inclusive).

The review api resource now has a fetchMany function, which takes a list of
review ids, and returns a list of reviews, sorted in the same order.