JSON API: Add jsonapi.from_parameter() helper
This function can be used to process a query parameter whose value (if
specified) should identify some supported resource. It is similar to
jsonapi.deduce(), but won't look at the context, and takes the query
parameter name as an argument.

Implement this for various resource classes.

Where this makes sense, parameters are interpreted as an id if valid
integers, and otherwise as a name or, in the case of commits, as a
commit reference (e.g. SHA-1.)

By using this helper in Changesets.multiple(), its interpretation of the
commit parameters (from, to and commit) is now extended to also support
commit ids.