JSON API: Add resource skipping mechanism
This mechanism lets a resource class's json() function raise an exception
if the value provided should not be included in the output. Typically,
its single() or multiple() wouldn't have produced the value in the first
place, but in some cases it might be more convenient to detect it in
json(). Also, the value could be linked by another resource, which did not
know whether to include it.

IOW, I expect this to be used sparingly, but it will be useful in some
cases.