Add api.changeset.Changeset.contributing_commits attribute
This attribute returns a set of commits that contribute to the changes
in the changeset, or None if such a set can not be accurately determined.

Essentially, it returns None if the changeset's `from_commit` is not an
ancestor of its `to_commit`, or if the Git command

  git rev-list ^from_commit to_commit

returns any commits that `from_commit` is not an ancestor of. The latter
in practice means that the changeset includes a merge commit that merges
in a branch with a branch point that `from_commit` is not an ancestor of.

Also add api.commitset.calculateFromRange(), which actually implements
this logic, and is used to implement the attribute.

The set of commits is also exposed by the JSON API's changesets resource.