Add `criticctl interactive` command
This will start a session (setting 'critic' and 'db' locals), optionally
impersonate a named user (if the --user/-u argument is specified) and then
call IPython.embed(), thus starting an interactive Python session.

This works both in production installs and in a quick-started instance.
In case of the latter, `criticctl` is installed in $(state_dir)/bin/. IOW,
if Critic was started with the command

  $ python quickstart.py -s /tmp/critic

then the interactive session would be started by running

  $ /tmp/critic/bin/criticctl interactive

On a system with a production install, the `criticctl` utility would
typically have been installed as `/usr/bin/criticctl`.

IPython must be installed manually. On Debian/Ubuntu, the package name is
`ipython`.