API: Make APIObject.cached() better, and add APIObject.cachedMany()
The APIObject.cached() helper (used as a decorator on fetch() functions)
now supports an argument that is the class of exception that should be
raised when an invalid id is used. When supplied, the fetch() function is
assumed to return the iterator returned by APIObject.make(), and
APIObject.cached() calls next(), catches StopIteration and raises the
"invalid id" exception.

The new APIObject.cachedMany() functions works similarly, but wraps a
typical fetchMany(). It looks up each id in the cache, and fetches any
that weren't found by calling the wrapped fetchMany(), and then returns
a list of objects in the same order as the argument list of ids. If any
ids weren't found, it raises an error of a class that must be supplied
as an argument to APIObject.cachedMany().