API: Improve refresh of cached objects on transaction commit/rollback
Instead of refreshing each cached object individually (and unconditionally
regardless of what the transaction did), do the refresh per class of
cached object, for all cached objects of the type at once. Also use the
set of tables modified by the transaction to determine whether a class of
object needs to refresh at all.

As part of this, the first-level cache key is changed from wrapper class
to implementation class. This is internal to the API, though, so doesn't
affect users.