Introduce new JSON based syntax highlight cache format
The old HTML based cache format is pretty awkward for the new code which
wants a higher level representation of the highlighted source code, and
therefore currently parses the HTML syntax highlighting server-side.

In the new format, each line in a cache file is a JSON encoded array of
items, where each item is either a string (representing a part of the
line that is not highlighted, e.g. white-space) or an array containing
two items, the token type as an integer and the token as a string.

The old and new formats co-exist, because the new-format cache files are
named "X.json" ("X.json.bz2") where the old-format files are named "X".