API / JSON API: Redesign changesets and filediffs
Remove the low-level chunks from the API (for now.) They are mostly not
useful, since the Filediff API exposes largely the same information.

Also have the basic structure of a changeset referencing many file changes
that resolve to corresponding file diffs. This moves file changes back as
a central part of the API.

Also avoid using file ids and paths directly, instead referencing
api.file.File objects. In the JSON API, "file" fields are now file ids,
but since they're actually api.file.File objects, "include=files" returns
all the relevant id <=> path mappings. This will avoid duplication of
paths in different JSON structures, while still always making them
available when needed.

Also request syntax highlighting of file versions eagerly (but
asynchronusly) and raise an `api.filediff.FilediffDelayed` exception if
the syntax highlighting is not available when needed.

This means we avoid blocking on syntax highlighting, and also vastly
improves performance, since we'll request all necessary highlighting at
once, thus allowing the syntax highlighting service to parallelize the
work. On the client-side, this will typically be handled by exactly the
same error handling that handles the whole changeset being unavailable.

New diff line format
--------------------
Instead of having two lists of parts, have just one, where each part is
either deleted, meaning it should be on the old side only, inserted, in
which case it should be on the new side only, or neither, in which case
it should be on both sides.

To reconstruct the old and new sides of a modified line from this single
list, one can simply do

  old_line = "".join(part.content
                     for part in line.content
                     if part.state != "i")
  new_line = "".join(part.content
                     for part in line.content
                     if part.state != "d")

The advantage is that this is more efficient for context lines, and lines
that are "lightly" modified (i.e. consist mostly of unchanged tokens,)
since the unchanged parts of the line will not be duplicated.

Also take advantage of the new JSON based syntax highlighting format, and
thus avoid "parsing" HTML.