JSON API: Support "output_format=static" query parameter
This changes the output format for single resources to be the same as for
multiple resources:

  {
    "resource_class": [{
      // the resource
    }]
  }

This is convenient for API users that load a set of resources using the
path

  /api/v1/<resource_class>/A,B,C

since it avoids "set of one" becoming a special case with differently
structured output.

I may also be useful for a generic response handler that maps each
resource class to some specific handler, since the name of the resource
class is always there in the response.