API / JSON API: Support automatic changesets in review context
In the context of a review, it's useful to be able to load certain
"automatic" changesets, like "all changes", "all reviewable changes"
and "all pending (reviewable) changes".

This implementation mostly reuses the old legacy implementation of
this feature, which is slightly extended to support "everything" as
a filter mode.

Also rename the changeset id argument to `api.changeset.fetch()` from
`id` to `changeset_id`. This matches how most locals containing an
id value are named (and avoids shadowing the built-in `id()`.)

The old UI's version of this feature also typically filters the list
of files, rather than just computing an automatic range of commits.
This is not done here, to keep things simple. It should probably be
done. Also, the "files" automatic mode (which is just a manual filter
of the files list for a given changeset) is not included. Not sure if
it is really that meaningful to have.