Add Critic version life-cycle support
Essentially, add a small JSON file (installation/lifecycle.json) that
primarily defines whether the current commit is a stable or unstable
version (i.e. on a stable or unstable branch.)

In install.py, warn when installing an unstable version (but the default
choice is to go ahead.)

In upgrade.py, warn when upgrading to an unstable version from a stable
one (including one from master before this commit.) The default choice is
to not go ahead with the upgrade. When upgrading from an unstable version
to a stable one, this is noted, but no questions are asked, as this is not
something that would be easy to do by mistake. (Such an "upgrade" is
likely a significant downgrade, though, and those are generally not really
supported, since migrations are one-way.)

This commit will be the branch point for the 'stable/1' branch. The next
commit on master will update redefine master as an unstable development
branch.