Disable testing for all instances
This lets us do the conversion in many steps instead of one big one,
without angry red test results, and will eventually let us land the
branch with lots of broken intermediate commits.

Broken intermediate commits is typically something we don't want in
this repository, but if there ever was an exception to this rule,
this would be it, I think.