Support upgrade to Python 3.x Critic
Specificallty, use absolute (or rather explicitly relative) imports in
the installed configuration module, so that it can also be imported in a
Python 3.x interpreter, which is required to run upgrade.py successfully
with a Python 3.x interpreter, which in turn is how it must be run when
upgrading to a future Python 3.x Critic commit.

Also add a (dummy) configuration setting that simply indicates (with its
mere presence) that the installation supports being upgraded to Python 3.