Critic 2.0: Major rewrite of everything
Critic is now a proper Python package, meant to be installed in a virtual
environment using `pip install`. The old installation scripts and
infrastructure are now obsolete. The src/ directory is renamed critic/ to
match Python packaging conventions. A single top-level package `critic`
is installed, and all imports in the code base are made relative.

The criticctl command is extended to support performing various "tasks",
such as performing the basic system installation (create system user/group,
create the database, and write the basic system configuration), selecting
and configuring WSGI container and HTTP frontend, and performing required
migrations post-upgrade. Due to its increased scope, the implementation is
refactored and modularized, moving each command and task into its own file.

Almost all system settings are now stored in the database. The only
configuration stored separately is that which is required to access the
database: system username/groupname, system identity, and database
connection parameters. All of that information is very easy to recreate,
meaning effectively the database and Git repositories is really the only
thing that needs to be persistent.

The whole code-base has been infected by asyncio, since all I/O (database
access, network communication, IPC, et c.) is done in an asynchronous
manner. Very many functions (and properties) now return coroutines.

Pre-commit hooks are added to ensure basic code quality. The `flake8`
style checked and linter is most significantly used. Correctness checks
are enabled for all source files, but many currently have errors. They are
to be fixed as they are touched. Style checks are limited to a subset of
the code. This is intended to be extended over time, but some legacy code
(old front-end, perhaps) may never be fixed.

Critic can now be run as a set of micro-services. Docker images can be
built using Dockerfiles in docker/dockerfiles/, and there is an example
docker-compose.yaml file. Tests can be run against a system running in
docker-compose.