Use real manifest path for extension snapshots
When extension.getManifest() is asked for manifest of a given sha1, it
reads the manifest from the snapshot directory. It then synthesises an
invalid path, for some reason. This makes the created manifest unusable
for actually running extensions.

But the extension has a snapshot on disk from which this manifest was
read! So return the install_path of that snapshot.

This fixes ability to run non-LIVE versions of filterhook role, which
uses getManifest to avoid reading the manifest too often.

Fixes: #138